+++ sample/src/androidMain/kotlin/com/github/panpf/zoomimage/sample/ui/examples/BaseZoomImageViewFragment.kt
@@ -203,6 +203,9 @@ abstract class BaseZoomImageViewFragment<ZOOM_VIEW : ZoomImageView> :
                     tileAnimationSpecState.value =
                         if (it) TileAnimationSpec.Default else TileAnimationSpec.None
                 }
+                appSettings.tileMemoryCache.collectWithLifecycle(viewLifecycleOwner) {
+                    disabledTileBitmapCacheState.value = !it
+                }
                 appSettings.pausedContinuousTransformTypes.collectWithLifecycle(viewLifecycleOwner) {
                     pausedContinuousTransformTypesState.value = it
                 }