+++ zoomimage-view-sketch-core/src/main/kotlin/com/github/panpf/zoomimage/SketchZoomImageView.kt
@@ -21,7 +21,6 @@ import android.graphics.drawable.Animatable
 import android.graphics.drawable.Drawable
 import android.util.AttributeSet
 import com.github.panpf.sketch.Sketch
-import com.github.panpf.sketch.cache.CachePolicy
 import com.github.panpf.sketch.imageResult
 import com.github.panpf.sketch.request.ImageResult
 import com.github.panpf.sketch.util.SketchUtils
@@ -92,8 +91,6 @@ open class SketchZoomImageView @JvmOverloads constructor(
                 if (tileBitmapCacheState.value == null) {
                     tileBitmapCacheState.value = SketchTileBitmapCache(sketch)
                 }
-                disabledTileBitmapCacheState.value =
-                    result.request.memoryCachePolicy != CachePolicy.ENABLED    // TODO remove
                 setImageSource(newImageSource(sketch, result))
             }
         }