+++ CHANGELOG.md
@@ -5,6 +5,8 @@ Translations: [简体中文](CHANGELOG_zh.md)
 ## new
 
 * change: The `contentSize` properties of ZoomableState and ZoomableEngine now no longer return `containerSize` when empty
+* change: Decisions about whether to disable Tile's memory cache are no longer based on the memory
+  cache setting of the thumbnail's image request
 
 ## 1.1.0-alpha06